refactor(sys): always use integer types from arch/types.h
Use those instead of primitive types such as short, int or long.

+++ sys/include/interrupt.h
@@ -28,10 +28,10 @@ un *interruptargs(void);
 /* Sets drv to the driver of the device which caused the interrupt and returns
    0 normally and returns -1 if the interrupt type is not HARDWARE or the
    device is unknown. */
-int interruptdriver(struct driver *drv);
+u8 interruptdriver(struct driver *drv);
 
 /* Returns 1 if interrupts are enabled and 0 otherwise. */
-int interruptsenabled(void);
+u8 interruptsenabled(void);
 
 /* Returns the interrupt type code. */
 u8 interrupttype(void);
